pyproject.toml
[project]
name = "arcreel"
version = "0.1.0"
description = "AI video generation workspace"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.1.29",
    "ffmpeg-python>=0.2.0",
    "fastapi>=0.109.0",
    "google-genai>=0.5.0",
    "Pillow>=10.0.0",
    "pydantic>=2.6.0",
    "PyYAML>=6.0.1",
    "pytest>=8.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "uvicorn>=0.27.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["lib", "webui"]
omit = [".claude/skills/*"]

[tool.coverage.report]
show_missing = true
skip_empty = true
